FIle generation fixed and files regenerated for testing

+++ PY Files/Generation.py
@@ -5,20 +5,22 @@ from Database import *
 from Subscriptions import *
 
 out = widgets.Output()
-button = widgets.Button(description='Generate files',disabled=False,button_style='',tooltip='Click me',icon='check')
+button = widgets.Button(description='Generate files', disabled=False, button_style='', tooltip='Click me', icon='check')
 sub_text = widgets.Text(value='100', placeholder='', description='Ne Subs', disabled=False)
 book_text = widgets.Text(value='50', placeholder='e.g. coab', description='N Bookings', disabled=False)
 rent_text = widgets.Text(value='0.5', placeholder='e.g. coab', description='Rental Chance', disabled=False)
-renting_text = widgets.HTML(value="",placeholder='',description='')
+renting_text = widgets.HTML(value="", placeholder='', description='')
 
 generation_widget = VBox([sub_text, book_text, rent_text, button, renting_text, out], layout=Layout(display='flex'))
+
 
 def g_subscriptions(n_subs):
           subscriptions = {}
           for i in range(n_subs):
                     start_date, end_date = g_dates()
-                    subscriptions[g_user()] = {'SubscriptionType': " " + g_sub_type(),'StartDate': " " +  start_date,'EndDate': " " +  end_date}
+                    subscriptions[g_user()] = {'SubscriptionType': " " + g_sub_type(), 'StartDate': " " + start_date, 'EndDate': " " + end_date}
           write_dict_to(subscriptions, "Subscriptions.txt")
+
 
 def g_bookings(n_bookings):
           bookings = {}
@@ -26,6 +28,7 @@ def g_bookings(n_bookings):
                     user_id = random.choice(list(load_subscriptions().keys()))
                     bookings[user_id] = {'Date': " " + g_dates()[0], 'Time': " " + g_time(), 'Guests': " " + g_guests()}
           write_dict_to(bookings, "Booking.txt")
+
 
 def g_rentals(rental_chance):
           Games = loadGameSearch(["digital", "board"])
@@ -41,7 +44,8 @@ def g_rentals(rental_chance):
 def on_enter(b):
           g_subscriptions(int(sub_text.value))
           g_bookings(int(book_text.value))
-          g_rentals(int(rent_text.value))
+          g_rentals(float(rent_text.value))
           renting_text.value = "<h3>Genration Successful</h3>"
+
 
 button.on_click(on_enter)
